[REFACTOR] move StatEffect to dedicated module
