+++ backend/plugins/foes/_base.py
@@ -7,8 +7,8 @@ from dataclasses import fields
 import logging
 
 from autofighter.character import CharacterType
+from autofighter.stat_effect import StatEffect
 from autofighter.stats import BUS
-from autofighter.stats import StatEffect
 from autofighter.stats import Stats
 from plugins.damage_types import random_damage_type
 from plugins.damage_types._base import DamageTypeBase
