+++ backend/plugins/passives/becca_menagerie_bond.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass
 from typing import TYPE_CHECKING
 from typing import ClassVar
 
-from autofighter.stats import StatEffect
+from autofighter.stat_effect import StatEffect
 from autofighter.summons.manager import SummonManager
 from plugins.damage_types import load_damage_type
