+++ backend/plugins/passives/carly_guardians_aegis.py
@@ -4,7 +4,7 @@ from typing import ClassVar
 from typing import Optional
 from weakref import WeakSet
 
-from autofighter.stats import StatEffect
+from autofighter.stat_effect import StatEffect
 
 if TYPE_CHECKING:
     from autofighter.stats import Stats
