+++ backend/plugins/passives/lady_fire_and_ice_duality_engine.py
@@ -3,7 +3,7 @@ from typing import TYPE_CHECKING
 from typing import ClassVar
 from typing import Iterable
 
-from autofighter.stats import StatEffect
+from autofighter.stat_effect import StatEffect
 
 if TYPE_CHECKING:
     from autofighter.stats import Stats
