+++ backend/plugins/passives/luna_lunar_reservoir.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass
 from typing import TYPE_CHECKING
 from typing import ClassVar
 
-from autofighter.stats import StatEffect
+from autofighter.stat_effect import StatEffect
 
 if TYPE_CHECKING:
     from autofighter.stats import Stats
