+++ backend/plugins/passives/mimic_player_copy.py
@@ -2,8 +2,8 @@ from dataclasses import dataclass
 from typing import TYPE_CHECKING
 from typing import ClassVar
 
+from autofighter.stat_effect import StatEffect
 from autofighter.stats import BUS
-from autofighter.stats import StatEffect
 
 if TYPE_CHECKING:
     from autofighter.stats import Stats
